Creating the Workers Module Moving MessagesComponent into Shared
